# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Book:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        author:
          type: string
          nullable: true
        isbn:
          type: string
          nullable: true
        language:
          type: string
        level:
          type: string
          nullable: true
        coverImageUrl:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - language
        - isActive
        - createdAt
        - updatedAt
    CreateBookRequest:
      type: object
      properties:
        title:
          type: string
        author:
          type: string
        isbn:
          type: string
        language:
          type: string
        level:
          type: string
        coverImageUrl:
          type: string
        description:
          type: string
      required:
        - title
    UpdateBookRequest:
      type: object
      properties:
        title:
          type: string
        author:
          type: string
        isbn:
          type: string
        language:
          type: string
        level:
          type: string
        coverImageUrl:
          type: string
        description:
          type: string
        isActive:
          type: boolean
...
    DigitizationRequest:
      type: object
      properties:
        id:
          type: integer
        institutionId:
          type: integer
        requestedBy:
          type: integer
        bookTitle:
          type: string
        bookAuthor:
          type: string
          nullable: true
        bookIsbn:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, in_progress, completed, rejected]
        notes:
          type: string
          nullable: true
        adminNotes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        institution:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
        requester:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
            email:
              type: string
      required:
        - id
        - institutionId
        - requestedBy
        - bookTitle
        - status
        - createdAt
        - updatedAt
    CreateDigitizationRequest:
      type: object
      properties:
        bookTitle:
          type: string
        bookAuthor:
          type: string
        bookIsbn:
          type: string
        notes:
          type: string
      required:
        - bookTitle
    UpdateDigitizationRequest:
      type: object
      properties:
        status:
          type: string
          enum: [pending, in_progress, completed, rejected]
        adminNotes:
          type: string
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
paths:
...
  /books:
    get:
      operationId: listBooks
      tags: [books]
      summary: List all active books
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Book"
    post:
      operationId: createBook
      tags: [books]
      summary: Add a book to the catalog
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookRequest"
      responses:
        "201":
          description: Book created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
  /books/{id}:
    get:
      operationId: getBook
      tags: [books]
      summary: Get book by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Book details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      operationId: updateBook
      tags: [books]
      summary: Update a book
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookRequest"
      responses:
        "200":
          description: Book updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteBook
      tags: [books]
      summary: Delete a book
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Book deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /digitization-requests:
    get:
      operationId: listDigitizationRequests
      tags: [digitization]
      summary: List digitization requests
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of digitization requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DigitizationRequest"
    post:
      operationId: createDigitizationRequest
      tags: [digitization]
      summary: Submit a digitization request
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDigitizationRequest"
      responses:
        "201":
          description: Request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DigitizationRequest"
  /digitization-requests/{id}:
    put:
      operationId: updateDigitizationRequest
      tags: [digitization]
      summary: Update digitization request status (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDigitizationRequest"
      responses:
        "200":
          description: Request updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DigitizationRequest"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"